Update PerformanceExplorer.java
In Windows OS, running PerformanceExplorer would encounter a problem; for file separator in Windows is "\" while in Vivado Tcl is “/”. Therefore, the Tcl script created by PerformanceExplorer in Windows OS would fail in Vivado. 
I suppose the simplest way to solve this problem would be making a replacement of all "\" in the Tcl script to "/".