Update ModuleInst.java for a more flexible connecting
When we were trying to connect: 1) one bit of a multi-bit port to another single-bit port, or 2) one bit of a multi-bit port to another bit of another multi-bit port (different indices), it seems that there was a restriction on the method connect(..). Therefore, I improved the method connect(...) to allow a more flexible connecting, which also allowed the previously mentioned connecting to be possible.