Update EDIFNetlist.java: migrateCellAndSubCells()
Sometimes, we could have different cells with the same EDIFName, e.g. when trying to generate pre-implemented modules, we can use different generic parameters to generate different entities, and thus different cells despite the same EDIFName (entity name). When we then try to use these modules (which contains different cells with the same EDIFName) in RapidWright, there would be a conflict. I propose a reasonable way to solve this problem, the way that instead of skipping migrating cells whose EDIFNames have already existed in the library, we add suffixes to the EDIFNames of new cells to distinguish, and thus all new cells could be migrated. However, I don't think it is the best way because in this way even an entirely same cell could also be migrated and renamed as a "different" one. Also, some space could also be wasted. Nevertheless, It could also be solved, I suppose, by change the method EDIFLibrary.containsCell() to check the library based on Value (EDIFCell) instead of Key (legalEDIFName), but I am not sure if this way could cause any other underlying problems...