Update EDIFNetlist.java: migrateCellAndSubCells()
Sometimes, we could have different cells with the same EDIFName, e.g. when trying to generate pre-implemented modules, we can use different generic parameters to generate different entities, and thus different cells despite the same EDIFName (entity name). When we then try to use these modules (which contains different cells with the same EDIFName) in RapidWright, there would be a conflict, and errors occur. I propose a reasonable way to solve this problem, the way that instead of skipping migrating cells whose EDIFNames have already existed in the library, we add suffixes to the EDIFNames of new cells to distinguish, and thus all new cells could be migrated. However, there is a drawback of this way that even an entirely same cell could also be migrated and renamed as a "different" one. I hope my pull request would help a little. Thanks.