Ignore TestCheckOpenFilesInstalled.test if outside of gradle
PR #397 modified the gradle environment such that every testcase is
automatically extended with methods to ensure the test closes all of the
files it opens. However when run outside of gradle (e.g. in an IDE) the9
testcase that checks to see if this extension is running obviously fails.
This commit adds a conditional to skip the test if run outside of
gradle.

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>