Ignore TestCheckOpenFilesInstalled.test if outside of gradle (#674)
* Ignore TestCheckOpenFilesInstalled.test if outside of gradle

PR #397 modified the gradle environment such that every testcase is
automatically extended with methods to ensure the test closes all of the
files it opens. However when run outside of gradle (e.g. in an IDE) the9
testcase that checks to see if this extension is running obviously fails.
This commit adds a conditional to skip the test if run outside of
gradle.

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>

* conditionally ignore TestCheckOpenFIlesInstalled.test

ignore the test with an @EnableIf statement that checks to see if the
test is being run in gradle. To check if the test is running in gradle
search all system properties for any mention of "gradle".

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>

* only run TestCheckOpenFilesInstalled.test if extensions enabled

instead of parsing the command line check to see if the system property
that enables the CheckOpenFilesExtension is set

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>

* only run TestCheckOpenFilesInstalled.test if extensions enabled

pedantic change

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>

* Extension to decide whether to skip, based on ExtensionContext

Signed-off-by: Eddie Hung <eddie.hung@amd.com>

* Update test/shared/com/xilinx/rapidwright/support/CheckOpenFilesExtension.java

Signed-off-by: eddieh-xlnx <eddie.hung@amd.com>

---------

Signed-off-by: Zak Nafziger <zak.nafziger@amd.com>
Signed-off-by: Eddie Hung <eddie.hung@amd.com>
Signed-off-by: eddieh-xlnx <eddie.hung@amd.com>
Co-authored-by: Eddie Hung <eddie.hung@amd.com>